requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
streamlit==1.38.0
streamlit-extras
huggingface_hub==0.16.4
beautifulsoup4

# Embeddings + RAG
sentence-transformers==2.2.2
transformers==4.35.2
accelerate==0.22.0
torch==2.2.1
faiss-cpu==1.7.4
rank-bm25==0.2.2

# Parsing de documents
python-docx==0.8.11
docx2txt==0.8
pypdf==4.2.0
pyyaml==6.0.2

# Utilitaires
numpy==1.26.4
pandas==2.2.3
tqdm==4.66.5
requests==2.31.0

